moved selection summary code into separate file
