+++ src/js/components/SelectMultiple/SelectMultipleContainer.js
@@ -20,7 +20,7 @@ import { InfiniteScroll } from '../InfiniteScroll';
 import { Keyboard } from '../Keyboard';
 import { Text } from '../Text';
 import { TextInput } from '../TextInput';
-
+import { SelectionSummary } from './SelectionSummary';
 import {
   StyledContainer,
   OptionsBox,
@@ -320,84 +320,6 @@ const SelectMultipleContainer = forwardRef(
       }
     }, [showA11yLimit]);
 
-    const selectedValuesDisabled = useCallback(() => {
-      let disabledSelected = 0;
-      for (let i = 0; i < allOptions.length; i += 1) {
-        if (
-          value.includes(getOptionValue(i, options, valueKey)) &&
-          checkDisabled(i, disabled, disabledKey, options, valueKey)
-        )
-          disabledSelected += 1;
-      }
-      if (value.length === disabledSelected) return true;
-      return false;
-    }, [value, allOptions, disabled, disabledKey, options, valueKey]);
-
-    const selectClearAll =
-      search === '' || search === undefined ? (
-        <>
-          <Box alignSelf="center">
-            <Text size="small">
-              {value.length === 0
-                ? `0 selected`
-                : `${value.length} selected of ${options.length}`}
-            </Text>
-          </Box>
-          {(options.length &&
-            (!limit || !(value.length === 0 && selectedValuesDisabled()))) >
-            0 && (
-            <Button
-              a11yTitle={
-                value.length === 0 || selectedValuesDisabled()
-                  ? `Select all ${options.length} options`
-                  : `${value.length} options selected. Clear all?`
-              }
-              label={
-                value.length === 0 || selectedValuesDisabled()
-                  ? 'Select All'
-                  : 'Clear All'
-              }
-              onClick={(event) => {
-                const selectAll =
-                  value.length === 0 || selectedValuesDisabled();
-                if (onChange) {
-                  const nextSelected = options.filter((i, index) =>
-                    selectAll
-                      ? !checkDisabled(
-                          index,
-                          disabled,
-                          disabledKey,
-                          options,
-                          valueKey,
-                        ) || isSelected(index)
-                      : checkDisabled(
-                          index,
-                          disabled,
-                          disabledKey,
-                          options,
-                          valueKey,
-                        ) && isSelected(index),
-                  );
-                  const nextValue = nextSelected.map((i) =>
-                    valueKey && valueKey.reduce ? applyKey(i, valueKey) : i,
-                  );
-                  onChange(event, {
-                    option: options,
-                    value: nextValue,
-                    selected: nextSelected,
-                  });
-                }
-                if (limit && !selectAll) setActiveIndex(0);
-              }}
-              onFocus={() => setActiveIndex(-1)}
-              ref={clearRef}
-            />
-          )}
-        </>
-      ) : (
-        <Text size="small">{`${value.length} selected`}</Text>
-      );
-
     return (
       <Keyboard
         onEnter={onSelectOption}
@@ -428,7 +350,20 @@ const SelectMultipleContainer = forwardRef(
                 gap="small"
                 fill="horizontal"
               >
-                {selectClearAll}
+                <SelectionSummary
+                  allOptions={allOptions}
+                  clearRef={clearRef}
+                  disabled={disabled}
+                  disabledKey={disabledKey}
+                  isSelected={isSelected}
+                  limit={limit}
+                  onChange={onChange}
+                  options={options}
+                  search={search}
+                  setActiveIndex={setActiveIndex}
+                  value={value}
+                  valueKey={valueKey}
+                />
               </Box>
               <Button
                 icon={<FormUp />}
@@ -445,7 +380,20 @@ const SelectMultipleContainer = forwardRef(
               gap="small"
               flex={false}
             >
-              {selectClearAll}
+              <SelectionSummary
+                allOptions={allOptions}
+                clearRef={clearRef}
+                disabled={disabled}
+                disabledKey={disabledKey}
+                isSelected={isSelected}
+                limit={limit}
+                onChange={onChange}
+                options={options}
+                search={search}
+                setActiveIndex={setActiveIndex}
+                value={value}
+                valueKey={valueKey}
+              />
             </Box>
           )}
